requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==6.0.0

# Data processing
pandas==2.1.3
pyarrow==14.0.1

# Math/calculations
numpy==1.26.2

# Environment variables
python-dotenv==1.0.0

# Production server (optional, recommended for deployment)
gunicorn==22.0.0

# Logging and monitoring
colorama==0.4.6

# Progress bars (for scripts)
tqdm==4.66.3

# HTTP requests (for scripts)
requests==2.32.4